backend/requirements.txt
Django==6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
python-decouple==3.8
Pillow==12.1.0
gunicorn==22.0.0
whitenoise==6.6.0
celery==5.3.4
redis==5.0.1
